requirements.txt
# Core Data Processing
polars==0.20.3
numpy==1.26.3
scipy==1.12.0
statsmodels==0.14.1
scikit-learn==1.4.0
pandas==2.2.0

# Time Series
statsforecast==1.7.3

# Causality
dowhy==0.11.1

# Anomaly Detection
pyod==1.1.3

# LLM Interface
groq==1.17.9
pydantic==2.5.3
python-dotenv==1.0.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2
kaleido==0.2.1

# Reporting
weasyprint==60.2
jinja2==3.1.3

# UI
streamlit==1.30.0

# GPU Support
torch==2.2.0
cupy-cuda12x==13.0.0

# Utilities
python-dateutil==2.8.2
